team activity form fetch events
- Critical Bug (when submitting event activity it will show error message)
